mcp-server/requirements.txt
# GPT Researcher dependencies
gpt-researcher>=0.12.16
python-dotenv

# MCP dependencies
mcp>=1.6.0
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0
fastapi_mcp

# Utility dependencies
loguru>=0.7.0
